==> Masters Thesis/python scripts/05_tensorflow.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers
# import torch
import seaborn as sns
import sklearn.metrics as metrics

df = pd.read_parquet("/Users/kenny/accure_local/e3dc/training-1.parquet")
# train = pd.read_parquet("/Users/kenny/accure_local/senec/test.parquet")
df.dropna(inplace=True)

normalizer = tf.keras.layers.Normalization(axis=-1)
normalizer.adapt(np.array(train_features))

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
x = df.drop(["accure_id","dsoh","soh"],axis=1).values
y = df["dsoh"].values
x_train, x_test, y_train, y_test = train_test_split(
    x, y, test_size=0.25, random_state=12)
sc = StandardScaler()
x_train = sc.fit_transform(x_train)
x_test = sc.transform(x_test)

from tensorflow.keras.layers import Input, Dense, Activation,Dropout
from tensorflow.keras.models import Model
input_layer = Input(shape=(x.shape[1],))
dense_layer_1 = Dense(100, activation='relu')(input_layer)
dense_layer_2 = Dense(50, activation='relu')(dense_layer_1)
dense_layer_3 = Dense(25, activation='relu')(dense_layer_2)
output = Dense(1)(dense_layer_3)

model = Model(inputs=input_layer, outputs=output)
model.compile(loss="mean_squared_error" , optimizer="adam", metrics=["mean_squared_error"])

history = model.fit(x_train, y_train, batch_size=2, epochs=100, verbose=1, validation_split=0.2)

pred_train = model.predict(x_train)
print(np.sqrt(metrics.mean_squared_error(y_train,pred_train)))

pred = model.predict(x_test)
print("RMSE: %.6f %%" % (np.sqrt(metrics.mean_squared_error(y_test, pred))))
